+++ data/invoice_items_fixture.csv
@@ -1,0 +1,31 @@
+id,item_id,invoice_id,quantity,unit_price,created_at,updated_at
+1,263519844,1,5,13635,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+2,263454779,1,9,23324,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+3,263451719,1,8,34873,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+4,263542298,1,3,2196,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+5,263515158,1,7,79140,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+6,263539664,1,5,52100,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+7,263563764,1,4,66747,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+8,263432817,1,6,76941,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+9,263529264,2,6,29973,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+10,263523644,2,4,1859,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+11,263532898,2,3,30949,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+12,263438971,2,8,31099,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+13,263553176,3,4,78660,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+14,263529264,3,7,4264,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+15,263533242,3,5,72018,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+16,263526970,3,9,72018,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+17,263517394,3,5,46383,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+18,263520800,3,5,78660,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+19,263509232,3,8,48718,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+20,263395237,3,9,72018,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+21,263566314,4,2,42203,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+22,263546142,4,3,37333,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+23,263422693,5,1,66412,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+24,263557908,5,10,49121,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+25,263452311,5,9,32346,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+26,263543136,5,10,73408,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+27,263553784,6,2,8342,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+28,263547824,6,3,48795,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+29,263547780,6,2,22098,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
+30,263543430,6,6,18222,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
