+++ data/invoices_fixture.csv
@@ -1,0 +1,31 @@
+id,customer_id,merchant_id,status,created_at,updated_at
+1,1,12335938,pending,2009-02-07,2014-03-15
+2,1,12334753,shipped,2012-11-23,2013-04-14
+3,1,12335955,shipped,2009-12-09,2010-07-10
+4,1,12334269,pending,2013-08-05,2014-06-06
+5,1,12335311,pending,2014-02-08,2014-07-22
+6,1,12334389,pending,2015-03-13,2015-04-05
+7,1,12335009,pending,2006-10-16,2013-12-24
+8,1,12337139,shipped,2003-11-07,2004-07-31
+9,2,12336965,shipped,2003-03-07,2008-10-09
+10,2,12334839,pending,2014-04-13,2015-01-20
+11,2,12334771,pending,2008-01-07,2014-05-30
+12,2,12336617,shipped,2014-05-06,2014-10-06
+13,3,12335955,shipped,2004-02-14,2010-03-26
+14,3,12335150,pending,2005-12-19,2010-08-18
+15,3,12336730,shipped,2012-07-21,2014-07-01
+16,5,12336113,shipped,2014-07-26,2014-08-25
+17,5,12334912,pending,2005-06-03,2015-07-01
+18,5,12334839,shipped,2001-12-13,2006-05-02
+19,5,12334372,shipped,2004-09-28,2012-08-28
+20,5,12336163,shipped,2010-09-17,2012-10-11
+21,5,12336069,shipped,2012-08-28,2016-01-19
+22,5,12335902,shipped,2013-12-26,2014-04-21
+23,5,12336652,pending,2012-12-20,2013-01-29
+24,6,12335989,shipped,2003-10-30,2008-06-22
+25,6,12334264,returned,2011-08-08,2015-07-21
+26,6,12335690,shipped,2008-04-04,2009-04-27
+27,6,12335319,shipped,2011-06-21,2013-07-22
+28,6,12336161,shipped,2013-05-26,2014-11-20
+29,7,12334861,shipped,2008-09-21,2010-10-21
+30,7,12334208,pending,2001-11-24,2009-11-13
