+++ lib/item.rb
@@ -1,21 +1,24 @@
 require 'pry'
 require "bigdecimal" 
 require "date"
+require_relative "../lib/data_access"
 
 class Item
-  attr_reader :id, :name, :description, :unit_price, :merchant_id, :created_at, :updated_at, :parent
+  include DataAccess
+
+  # attr_reader :id, :name, :description, :unit_price, :merchant_id, :created_at, :updated_at, :parent    ###Now done by DataAccess module
   
-  def initialize(data, parent=nil)
-    @id  = data[:id]
-    @name = data[:name]
-    @description = data[:description]
-    @unit_price =  data[:unit_price]
-    @merchant_id = data[:merchant_id]
-    @created_at  = data[:created_at]
-    @updated_at = data[:updated_at]
-    @parent = parent
-    #  binding.pry
-  end
+  # def initialize(data, parent=nil)
+  #   @id  = data[:id]
+  #   @name = data[:name]
+  #   @description = data[:description]
+  #   @unit_price =  data[:unit_price]
+  #   @merchant_id = data[:merchant_id]
+  #   @created_at  = data[:created_at]
+  #   @updated_at = data[:updated_at]
+  #   @parent = parent
+  #   #  binding.pry
+  # end
 
 # unit_price_to_dollars - returns the price of the item in dollars formatted as a Float
   def unit_price_to_dollars
