+++ lib/merchant.rb
@@ -1,11 +1,13 @@
+require_relative "../lib/data_access"
 class Merchant
-  attr_reader :id, :name, :parent
+  include DataAccess
+  # attr_reader :id, :name, :parent
   
-  def initialize(data, parent=nil)
-    @id  = data[:id]
-    @name = data[:name]
-    @parent = parent
-  end
+  # def initialize(data, parent=nil)
+  #   @id  = data[:id]
+  #   @name = data[:name]
+  #   @parent = parent
+  # end
 
   def items
     parent.parent.items.find_all_by_merchant_id(@id)
