+++ lib/merchant_repository.rb
@@ -1,16 +1,18 @@
 require_relative 'merchant'
+require_relative '../lib/data_access'
 
 class MerchantRepository
+  include DataAccess
   attr_reader :csv_file, :all, :parent
   
   def initialize(path, parent=nil)
     @csv_file = CSV.open(path, headers: true, header_converters: :symbol)
     @all = []
     @parent = parent
-    make_merchants
+    populate_repo
   end
 
-  def make_merchants
+  def populate_repo
     csv_file.each do |row|
       merchant = Merchant.new({:id => row[:id].to_i,
         :name => row[:name]}, self)
@@ -19,20 +21,23 @@ class MerchantRepository
       end
   end
 
-  def find_by_id(id)
-    all.find{ |merchant| merchant.id == id }
-  end
 
-  def find_by_name(name)
-    all.find{ |merchant| merchant.name.downcase == name.downcase }
-  end
+#######################
+  # def find_by_id(id)
+  #   all.find{ |merchant| merchant.id == id }
+  # end
 
-  def find_all_by_name(name_frag)
-     all.select{ |merchant| merchant.name.downcase.include?(name_frag.downcase) }
-  end
+  # def find_by_name(name)
+  #   all.find{ |merchant| merchant.name.downcase == name.downcase }
+  # end
+
+  # def find_all_by_name(name_frag)
+  #    all.select{ |merchant| merchant.name.downcase.include?(name_frag.downcase) }
+  # end
 
   def inspect
     "#<#{self.class} #{@merchants.size} rows>"
   end
+###########################
 end
