+++ test/test_helper.rb
@@ -3,6 +3,7 @@
 require 'minitest/autorun'
 require 'minitest/pride'
 require './lib/sales_engine'
+require './lib/data_access'
 require 'bigdecimal'
 require 'csv'
 require 'date'
